+++ vrc-get-vpm/src/environment.rs
@@ -1,7 +1,6 @@
 use crate::repo_holder::RepoHolder;
 use crate::repository::{RemotePackages, RemoteRepository};
 use crate::structs::package::PackageJson;
-use crate::structs::repo_cache::LocalCachedRepository;
 use crate::structs::setting::UserRepoSetting;
 use crate::traits::PackageCollection;
 use crate::utils::{JsonMapExt, PathBufExt};
@@ -21,6 +20,7 @@ use std::path::{Path, PathBuf};
 use std::{env, fmt, io};
 use tokio::fs::{create_dir_all, remove_file, File};
 use tokio::io::AsyncWriteExt;
+use crate::repository::local::LocalCachedRepository;
 
 /// This struct holds global state (will be saved on %LOCALAPPDATA% of VPM.
 #[derive(Debug)]
