+++ vrc-get-vpm/src/repository/mod.rs
@@ -1,3 +1,5 @@
 mod remote;
+pub(crate) mod local;
+
 pub use remote::RemotePackages;
 pub use remote::RemoteRepository;
